Improve error and disconnection handling.
Underlying socket errors are now always passed to the delegate, so that if the connection is unexpectedly closed you receive an error.

webSocketDidClose: is not called until the socket has actually closed. (Previously it was called in willDisconnectWithError:, swallowing the error.)

Calling -disconnect disconnects immediately, instead of after any pending reads or writes. This seems more in line with what a library user would expect.